+++ src/app/datagrid/datagrid-standard-formatter.demo.ts
@@ -61,6 +61,20 @@ export class DataGridStandardFormatterDemoComponent implements OnInit {
       showPercentText: true,
     };
 
+    // Add `sortFunction` and `sortable: true` to column `orderDate`
+    if (PAGING_COLUMNS && Array.isArray(PAGING_COLUMNS)) {
+      PAGING_COLUMNS.forEach((col) => {
+        if (col.id === 'orderDate') {
+          col.sortable = true;
+          col.sortFunction = function (value: any) {
+            const formatDateStr = Soho.Locale.formatDate(value, { pattern: 'M/d/yyyy' });
+            const time = Soho.Locale.parseDate(formatDateStr).getTime();
+            return time;
+          };
+        }
+      });
+    }
+
     const columns = [ ...PAGING_COLUMNS, statusColumn, ratedColumn ];
 
     return {