+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 6.3.0 Fixes
 
 - `[Datagrid]` Added `enableTooltips` Input to datagrid. `BTHH` ([#674](https://github.com/infor-design/enterprise-ng/pull/674))
+- `[FileUpload]` Fixed duplicate change event. `TJM` ([#672](https://github.com/infor-design/enterprise-ng/pull/672))
 
 ## v6.2.0